this fixes bug #514653. Short explanation:
In this particular case (as given in the sample code), the sexpr_eval function tried to compile code which was invalid
from the Pike point of view:

mixed foo(){ return +;}

which caused a backtrace as the compile_string function throws an exception which wasn't caught by sexpr_eval.
do_output_tag behaves correctly by substituting an empty string for non-existant variables (which was the case here).
If you want to make sure a math operation as above is compiled correctly, pass the 'zero' attribute to any output
tag involved. Using the sample from the original bug:

<formoutput zero="0"><set variable=GT expr="#GT# +
#Z#"></formoutput>
<formoutput><set variable=GT expr="#GT# -
1"></formoutput>
<formoutput>
<pre>
GT : #GT#
Z : #Z#
</pre>
</formoutput>

The above code shouldn't fail even with an unpatched server. Nevertheless, it was an obvious bug with sexpr_eval and it is
fixed now. Btw, Caudium 1.0 did exactly the same - the do_output_tag code doesn't differ in that respect.