ok, this was the wrong approach. fuck.
backed out all of dsn and a couple of other stuff probably. some fetures of
esmtp (the spec, not the mod:) need to be moved into a different subclass,
welcome to protocols._esmtp.async_client().

reason behind this is that stuff like dsn, size and a couple of others just
can'be implemented properly without knowing the whole message before there's
even an attempt to assemble what wil be communicated with the server (well,
maybe it *can* be done if one tries hard enough, but you wouldn't want to
touch that code ever, should it need be tweaked.

this _esmtp.client() here _probably_ works for straightforward connections,
but don't rely on it. async_client() is to come later this night, there
you'll have a transaction-ish interface for sending the mail. isn't
this word "transaction" just cool ? makes you think i actually know what i'm
doing ;P