session hash is generated on every request - no counter used anymore. It turns out that the time needed to generate the hash is minimal (Athlon XP 1700+ needs 4.9s to generate 100000 hashes with Pike 7.2 and Crypto.sha(), 4.3s using Pike 7.3 and Crypto.sha() and 4.3s using Pike 7.2 and Mhash.hash_sha() )
the session ID from a query (bookmarks, search engines etc.) is ignored when the referring site is not confirmed to be valid. By default valid referrer is the URL assigned to the virtual host the module lives in. This is done so that it's not possible to forge the session.